Hash Functions

Functions to calculate cryptographic hash of a message, with optionally a key for HMAC applications. For storing passwords, use password_store instead.

Usage

hash(buf, key = NULL, size = 32)

scrypt(buf, salt = raw(32), size = 32)

argon2(buf, salt = raw(16), size = 32)

shorthash(buf, key)

sha512(buf, key = NULL)

sha256(buf, key = NULL)

Arguments

buf

data to be hashed

key

key for HMAC hashing. Optional, except for in shorthash.

size

length of the output hash. Must be between 16 and 64 (recommended is 32)

salt

non-confidential random data to seed the algorithm

Details

The generic hash function is recommended for most applications. It uses dynamic length BLAKE2b where output size can be any value between 16 bytes (128bit) and 64 bytes (512bit).

The scrypt hash function is designed to be CPU and memory expensive to protect against brute force attacks. This algorithm is also used by the password_store function.

The argon2 hash function is also designed to be CPU and memory expensive to protect against brute force attacks. Argon2 is a password-hashing function that summarizes the state of the art in the design of memory-hard functions

The shorthash function is a special 8 byte (64 bit) hash based on SipHash-2-4. The output of this function is only 64 bits (8 bytes). It is useful for in e.g. Hash tables, but it should not be considered collision-resistant.

Hash functions can be used for HMAC by specifying a secret key. They key size for shorthash is 16 bytes, for sha256 it is 32 bytes and for sha512 it is 64 bytes. For hash the key size can be any value between 16 and 62, recommended is at least 32.

References

https://libsodium.gitbook.io/doc/hashing/generic_hashing

Examples

# Basic hashing
msg <- serialize(iris, NULL)
hash(msg)
#>  [1] 59 21 a7 79 8e 53 0b 60 ad 4e 53 04 15 15 df 38 86 f1 97 42 0d 3b 4a be 40
#> [26] f5 e7 d4 66 c2 19 8c
sha256(msg)
#>  [1] d7 d7 0e 07 79 9f 8e ba ef 70 83 35 4f f9 ec a0 5b 77 ce d4 71 1e 49 be b0
#> [26] 47 6e 10 8e 19 c8 88
sha512(msg)
#>  [1] 9b b2 b3 2e a7 ef 60 2c 31 76 69 90 02 1d 32 8c a1 e8 bb 77 b2 c6 00 0b 5e
#> [26] 78 73 3a f8 f8 17 1c 9e 3e ef 29 24 3b 7b 39 8c c3 62 84 b6 08 62 4f 0a 1f
#> [51] 72 c1 10 75 fe a5 11 76 66 d8 67 75 2e 0b
scrypt(msg)
#>  [1] 78 7e 51 07 44 e7 84 0e 57 bc d1 32 1e 72 0a 0b fd bc 5e 86 b2 0c 0d 64 c7
#> [26] 13 bb cb 11 d4 23 29

# Generate keys from passphrase
passphrase <- charToRaw("This is super secret")
key <- hash(passphrase)
shortkey <- hash(passphrase, size = 16)
longkey <- hash(passphrase, size = 64)

# HMAC (hashing with key)
hash(msg, key = key)
#>  [1] 63 7d 4f 86 ec f4 43 fc 5c 3d 9d 54 28 78 9b de 47 02 d7 01 0a 12 e8 73 0f
#> [26] 40 46 cd e0 e2 d9 b8
shorthash(msg, shortkey)
#> [1] 03 cd cb 4c cf 5e 8f 7e
sha256(msg, key = key)
#>  [1] 20 e3 5f 2b 73 85 89 87 03 6a 7b 22 d7 88 e6 cc 68 d8 8d 3a d7 1d b5 4e fd
#> [26] c0 e4 fd 38 c4 e8 33
sha512(msg, key = longkey)
#>  [1] 95 83 e5 3b f9 8c 8b 6c de 51 74 9f 7d cc 60 f0 ae 6a ae 76 b6 91 d9 bd 02
#> [26] 96 7f 80 99 74 94 f8 0d 0a 11 12 12 91 15 de 88 f2 b8 c5 39 b6 aa e6 91 2e
#> [51] 56 b7 a9 40 12 10 75 b8 64 00 4b 2a 8c db